3 hw fixes

+++ src/main/java/ok/dht/test/anikina/consistenthashing/ConsistentHashingImpl.java
@@ -13,17 +13,17 @@ import java.util.Set;
 public class ConsistentHashingImpl {
     private static final int VIRTUAL_NODES_COUNT = 3;
 
-    private final long[] hashes;
-    private final Map<Long, String> hashToShard = new HashMap<>();
+    private final int[] hashes;
+    private final Map<Integer, String> hashToShard = new HashMap<>();
 
     public ConsistentHashingImpl(List<String> clusterUrls) {
-        hashes = new long[clusterUrls.size() * VIRTUAL_NODES_COUNT];
+        hashes = new int[clusterUrls.size() * VIRTUAL_NODES_COUNT];
 
         for (int i = 0; i < clusterUrls.size(); i++) {
             String serverUrl = clusterUrls.get(i);
             for (int j = 0; j < VIRTUAL_NODES_COUNT; j++) {
                 String virtualNode = serverUrl + "_" + j;
-                long hash = hashForKey(virtualNode);
+                int hash = hashForKey(virtualNode);
                 hashes[i * VIRTUAL_NODES_COUNT + j] = hash;
                 hashToShard.put(hash, serverUrl);
             }
@@ -32,13 +32,13 @@ public class ConsistentHashingImpl {
         Arrays.sort(hashes);
     }
 
-    private long hashForKey(String key) {
+    private int hashForKey(String key) {
         byte[] keyBytes = Utf8.toBytes(key);
         return Hash.xxhash(keyBytes, 0, keyBytes.length);
     }
 
     public Set<String> getNodesByKey(String key, int replicas) {
-        long hash = hashForKey(key);
+        int hash = hashForKey(key);
 
         int shardIndex = Arrays.binarySearch(hashes, hash);
         if (shardIndex < 0) {
@@ -47,7 +47,7 @@ public class ConsistentHashingImpl {
 
         Set<String> nodes = new HashSet<>();
         for (int i = 0; i < hashes.length; i++) {
-            long currHash = hashes[(shardIndex + i) % hashes.length];
+            int currHash = hashes[(shardIndex + i) % hashes.length];
             nodes.add(hashToShard.get(currHash));
             if (nodes.size() == replicas) {
                 break;
